feat: base for backend of libsql implementation
