+++ packages/nc-gui/components/workspace/integrations/EditOrAdd.vue
@@ -18,16 +18,16 @@ const isEditOrAddIntegrationModalOpen = computed({
 
 // TODO TITOUAN FIND OUT WHY IS DOESNT WORK WHEN USING THE IMPORTED TYPE
 // this seems to not pick up the added LIBSQL key
-enum ClientType {
-  MYSQL = 'mysql2',
-  MSSQL = 'mssql',
-  PG = 'pg',
-  SQLITE = 'sqlite3',
-  VITESS = 'vitess',
-  SNOWFLAKE = 'snowflake',
-  DATABRICKS = 'databricks',
-  LIBSQL = 'libsql',
-}
+// enum ClientType {
+//   MYSQL = 'mysql2',
+//   MSSQL = 'mssql',
+//   PG = 'pg',
+//   SQLITE = 'sqlite3',
+//   VITESS = 'vitess',
+//   SNOWFLAKE = 'snowflake',
+//   DATABRICKS = 'databricks',
+//   LIBSQL = 'libsql',
+// }
 
 const connectionType = computed(() => {
   switch (
