+++ packages/nocodb-sdk/tsconfig.json
@@ -37,15 +37,16 @@
     // "experimentalDecorators": true /* Enables experimental support for ES7 decorators. */,
     // "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */,
 
-    "lib": ["es2017","dom"],
+    "lib": ["es2017", "dom"],
     "types": ["jest"],
     "typeRoots": ["node_modules/@types", "src/types"],
     "baseUrl": "./src",
     "paths": {
-      "~/*" : ["./*"],
-      "@/*" : ["./*"],
-      "src/*" : ["./*"]
-    }
+      "~/*": ["./*"],
+      "@/*": ["./*"],
+      "src/*": ["./*"]
+    },
+    "skipLibCheck": true
   },
   "include": ["src/**/*.ts"],
   "exclude": ["node_modules/**", "src/ee/**"],
